requirement.txt
asgiref==3.4.1
autopep8==2.0.0
blurhash-python==1.1.3
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.0.12
Django==3.2.15
django-cors-headers==3.10.1
django-crontab==0.7.1
et-xmlfile==1.1.0
gunicorn==20.1.0
idna==3.4
numpy==1.19.5
openpyxl==3.0.10
pandas==1.1.5
Pillow==8.4.0
psycopg2==2.7.7
pycodestyle==2.9.1
pycparser==2.21
python-dateutil==2.8.2
pytz==2022.2.1
requests==2.27.1
six==1.16.0
sqlparse==0.4.2
tomli==1.2.3
typing_extensions==4.1.1
urllib3==1.26.12
xlrd==1.2.0
